+++ runtime/vm/compiler/aot/precompiler.cc
@@ -871,17 +871,6 @@ void Precompiler::ProcessFunction(const Function& function) {
               function.token_pos().ToCString(),
               Function::KindToCString(function.kind()));
   }
-  if (function.SourceSize() >= FLAG_huge_method_cutoff_in_tokens) {
-    THR_Print(
-        "Warning: %s from %s is too large. Some optimizations have been "
-        "disabled, and the compiler might run out of memory. "
-        "Consider refactoring this code into smaller components.\n",
-        function.QualifiedUserVisibleNameCString(),
-        String::Handle(
-            Z, Library::Handle(Z, Class::Handle(Z, function.Owner()).library())
-                   .url())
-            .ToCString());
-  }
 
   ASSERT(!function.is_abstract());