+++ sdk/lib/_internal/vm/bin/secure_socket_patch.dart
@@ -33,12 +33,7 @@ class _SecureSocket extends _Socket implements SecureSocket {
   void renegotiate(
       {bool useSessionCache: true,
       bool requestClientCertificate: false,
-      bool requireClientCertificate: false}) {
-    _raw!.renegotiate(
-        useSessionCache: useSessionCache,
-        requestClientCertificate: requestClientCertificate,
-        requireClientCertificate: requireClientCertificate);
-  }
+      bool requireClientCertificate: false}) {}
 
   X509Certificate? get peerCertificate {
     if (_raw == null) {
@@ -164,10 +159,6 @@ class _SecureFilterImpl extends NativeFieldWrapperClass1
 
   @pragma("vm:external-name", "SecureSocket_GetSelectedProtocol")
   external String? selectedProtocol();
-
-  @pragma("vm:external-name", "SecureSocket_Renegotiate")
-  external void renegotiate(bool useSessionCache, bool requestClientCertificate,
-      bool requireClientCertificate);
 
   @pragma("vm:external-name", "SecureSocket_Init")
   external void init();